Move mock objects to separate files
