+++ spec/mastermind/game_spec.rb
@@ -1,70 +1,19 @@
 require 'spec_helper'
 require 'mastermind/game'
+require 'mastermind/mock/mock_ai'
+require 'mastermind/mock/mock_display'
 
 module Mastermind
-  class StubAI
-    attr_reader :next_guess_times_called
-
-    def initialize
-      @next_guess_times_called = 0
-    end
-    def next_guess(prev_exact_matches, prev_unexact_matches)
-      @next_guess_times_called += 1
-      "RRRR"
-    end
-  end
-
-  class StubDisplay
-
-    attr_accessor :display_welcome_times_called, 
-      :get_exact_matches_times_called, 
-      :get_unexact_matches_times_called,
-      :guess_message_times_called
-
-    def initialize
-      @display_welcome_times_called = 0
-      @guess_message_times_called = 0
-      @get_unexact_matches_times_called = 0
-      @get_exact_matches_times_called = 0
-    end
-
-    def display_welcome
-      @display_welcome_times_called += 1
-    end
-
-    def guess_message(guess, guess_number)
-      @guess_message_times_called += 1
-    end
-
-    def get_exact_matches
-      @get_exact_matches_times_called += 1
-      @exact_matches
-    end
-
-    def get_unexact_matches
-      @get_unexact_matches_times_called += 1
-      @unexact_matches
-    end
-
-    def set_exact_matches(number)
-      @exact_matches = number
-    end
-
-    def set_unexact_matches(number)
-      @unexact_matches = number
-    end
-  end
-
   describe Game do
-    let(:stub_ai) { StubAI.new }
-    let(:stub_display) { StubDisplay.new }
-    let(:game) { Game.new(stub_display) }
+    let(:mock_ai) { Mastermind::Mock::MockAI.new }
+    let(:mock_display) { Mastermind::Mock::MockDisplay.new }
+    let(:game) { Game.new(mock_display) }
 
     describe "#start" do
       it "sends a welcome message when the game starts" do
-        expect(stub_display.display_welcome_times_called).to eq(0)
+        expect(mock_display.display_welcome_times_called).to eq(0)
         game.start
-        expect(stub_display.display_welcome_times_called).to eq(1)
+        expect(mock_display.display_welcome_times_called).to eq(1)
       end
 
       it "starts with a guess counter at 0" do
@@ -91,33 +40,33 @@ module Mastermind
 
     describe "#play_turn" do
       it "asks the AI for a guess" do
-        game = Game.new(stub_display, stub_ai)
+        game = Game.new(mock_display, mock_ai)
 
         game.play_turn
-        expect(stub_ai.next_guess_times_called).to eq(1)
+        expect(mock_ai.next_guess_times_called).to eq(1)
       end
 
       it "calls display.guess_message" do
-        game = Game.new(stub_display, stub_ai)
+        game = Game.new(mock_display, mock_ai)
 
         game.play_turn
-        expect(stub_display.guess_message_times_called).to eq(1)
+        expect(mock_display.guess_message_times_called).to eq(1)
       end
 
       it "calls display.get_exact_matches" do
-        stub_display.set_exact_matches(0)
-        game = Game.new(stub_display, stub_ai)
+        mock_display.set_exact_matches(0)
+        game = Game.new(mock_display, mock_ai)
 
         game.play_turn
-        expect(stub_display.get_exact_matches_times_called).to eq(1)
+        expect(mock_display.get_exact_matches_times_called).to eq(1)
       end
 
       it "calls display.get_unexact_matches" do
-        stub_display.set_unexact_matches(0)
-        game = Game.new(stub_display, stub_ai)
+        mock_display.set_unexact_matches(0)
+        game = Game.new(mock_display, mock_ai)
 
         game.play_turn
-        expect(stub_display.get_unexact_matches_times_called).to eq(1)
+        expect(mock_display.get_unexact_matches_times_called).to eq(1)
       end
 
     end
